+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Fixed
 * Race condition on win32 in `ubelt.symlink`
+* Issue with `ubelt.memoize_method` where the method name and docstring were
+  not wrapped correctly.
 
 
 ## Version 1.1.2 - Released 2022-06-30